model category

+++ express/src/controllers/category.js
@@ -1,9 +1,9 @@
 const Category = require('./../models/category');
 
 //GET ALL ---> GET
-const getAll = async(req,res,next)=>{
+const getAll = async (req, res, next) => {
     try {
-        const rows = await Category.findAll();
+        const { rows } = await Category.findAll();
         res.status(200).json({ data: rows });
     } catch (error) {
         res.status(400).json({ error: error });
@@ -11,13 +11,13 @@ const getAll = async(req,res,next)=>{
     }
 }
 //GET BY ID ---> GET
-const getByID = async (req,res,next)=>{
+const getByID = async (req, res, next) => {
     const { id } = req.params;
     const args = {
         id: Number(id)
     }
     try {
-        const rows = await Category.findById(args);
+        const { rows } = await Category.findById(args);
         res.status(200).json({ data: rows });
     } catch (error) {
         res.status(400).json({ error: error });
@@ -25,10 +25,10 @@ const getByID = async (req,res,next)=>{
     }
 }
 //CREATE ---> POST
-const createCategory = async (req,res,next)=>{
+const createCategory = async (req, res, next) => {
     console.log(req.person);
-    if(!req.body.title || !req.body.body) {
-        return res.status(400).json({ message: "missing to enter data"});
+    if (!req.body.title || !req.body.body) {
+        return res.status(400).json({ message: "missing to enter data" });
     }
     const args = {
         person_id: req.person.id,
@@ -44,34 +44,46 @@ const createCategory = async (req,res,next)=>{
     }
 }
 //DELETE ---> DELETE
-const deleteCategory = async (req,res,next)=>{
-    if(!req.body.id) {
-        return res.status(400).json({ message: "missing to enter data"});
-    }
-    const args = {
-        id: Number(req.body.id),
+const deleteCategory = async (req, res, next) => {
+    if (!req.body.id) {
+        return res.status(400).json({ message: "missing to enter data" });
     }
     try {
-        await Category.deleteById(args);
-        res.status(200).json({ message: 'deleted' });
+        const args = {
+            id: req.body.id,
+            person_id: req.person.id,
+        }
+        const { rows } = await Category.personIDbyID(args);
+        //console.log(rows);
+        if (rows.length > 0 && rows[0]['PERSON_ID'] === args.person_id) {
+            await Category.deleteById(args);
+            return res.status(200).json({ message: 'deleted' });
+        }
+        res.status(200).json({ message: 'faltan permisos' });
     } catch (error) {
         res.status(400).json({ error: error });
         //next(error);
     }
 }
 //UPDATE ---> PUT
-const updateCategory = async (req,res,next)=>{
-    if(!req.body.title || !req.body.body || !req.body.id) {
-        return res.status(400).json({ message: "missing to enter data"});
-    }
-    const args = {
-        id: req.body.id,
-        title: req.body.title,
-        body: req.body.body
+const updateCategory = async (req, res, next) => {
+    if (!req.body.title || !req.body.body || !req.body.id) {
+        return res.status(400).json({ message: "missing to enter data" });
     }
     try {
-        await Category.updateById(args);
-        res.status(200).json({ message: 'update' });
+        const args = {
+            id: req.body.id,
+            person_id: req.person.id,
+            title: req.body.title,
+            body: req.body.body
+        }
+        const { rows } = await Category.personIDbyID(args);
+        //console.log(rows);
+        if (rows.length > 0 && rows[0]['PERSON_ID'] === args.person_id) {
+            await Category.updateById(args);
+            return res.status(200).json({ message: 'update' });
+        }
+        res.status(200).json({ message: 'faltan permisos' });
     } catch (error) {
         res.status(400).json({ error: error });
         //next(error);
